+++ src/fidesops/ops/api/v1/endpoints/drp_endpoints.py
@@ -66,7 +66,7 @@ async def create_drp_privacy_request(
     a corresponding Fidesops PrivacyRequest
     """
 
-    jwt_key: str = config.security.drp_jwt_secret
+    jwt_key: Optional[str] = config.security.drp_jwt_secret
     if jwt_key is None:
         raise HTTPException(
             status_code=HTTP_500_INTERNAL_SERVER_ERROR,