+++ src/fidesops/ops/core/config.py
@@ -25,7 +25,7 @@ logger = logging.getLogger(__name__)
 class FidesopsDatabaseSettings(DatabaseSettings):
     """Configuration settings for Postgres."""
 
-    ENABLED: bool = True
+    enabled: bool = True
 
     class Config:
         env_prefix = "FIDESOPS__DATABASE__"
@@ -35,9 +35,10 @@ class ExecutionSettings(FidesSettings):
     """Configuration settings for execution."""
 
     privacy_request_delay_timeout: int = 3600
-    task_retry_count: int
-    task_retry_delay: int  # In seconds
-    task_retry_backoff: int
+    # By default Fidesops will not retry graph nodes
+    task_retry_count: int = 0
+    task_retry_delay: int = 0  # In seconds
+    task_retry_backoff: int = 0
     subject_identity_verification_required: bool = False
     require_manual_request_approval: bool = False
     masking_strict: bool = True
@@ -77,7 +78,7 @@ class RedisSettings(FidesSettings):
             # If the whole URL is provided via the config, preference that
             return v
 
-        return f"redis://{quote_plus(values.get('user', ''))}:{quote_plus(values['password'])}@{values['host']}:{values['port']}/{values.get('db_index', '')}"
+        return f"redis://{quote_plus(values.get('user', ''))}:{quote_plus(values.get('password', ''))}@{values.get('host', '')}:{values.get('port', '')}/{values.get('db_index', '')}"
 
     class Config:
         env_prefix = "FIDESOPS__REDIS__"
@@ -118,15 +119,21 @@ class FidesopsSecuritySettings(SecuritySettings):
 class RootUserSettings(FidesSettings):
     """Configuration settings for Analytics variables."""
 
-    analytics_opt_out: Optional[bool]
-    analytics_id: Optional[str]
+    analytics_opt_out: Optional[bool] = True
+    analytics_id: Optional[str] = None
 
     @validator("analytics_id", pre=True)
-    def populate_analytics_id(cls, v: Optional[str]) -> str:
+    def populate_analytics_id(
+        cls,
+        v: Optional[str],
+        values: Dict[str, str],
+    ) -> Optional[str]:
         """
         Populates the appropriate value for analytics id based on config
         """
-        return v or cls.generate_and_store_client_id()
+        if not v and not values.get("analytics_opt_out"):
+            v = cls.generate_and_store_client_id()
+        return v
 
     @staticmethod
     def generate_and_store_client_id() -> str:
@@ -164,15 +171,29 @@ class FidesopsNotificationSettings(FidesSettings):
 class FidesopsConfig(FidesSettings):
     """Configuration variables for the FastAPI project"""
 
-    database: FidesopsDatabaseSettings
-    redis: RedisSettings
-    security: FidesopsSecuritySettings
-    execution: ExecutionSettings
-    root_user: RootUserSettings
-    admin_ui: AdminUiSettings
-    notifications: FidesopsNotificationSettings
+    # Pydantic doesn't initialise subsections automatically if
+    # only environment variables are provided at runtime. If the
+    # config subclass is instantiated with no args, Pydantic runs
+    # validation before loading in environment variables, which
+    # always fails if any config vars in the subsection are non-optional.
+    # Using the empty dict allows Python to load in the environment
+    # variables _before_ validating them against the Pydantic schema.
+    database: FidesopsDatabaseSettings = {}  # type: ignore
+    redis: RedisSettings = {}  # type: ignore
+    security: FidesopsSecuritySettings = {}  # type: ignore
+    execution: Optional[ExecutionSettings] = ExecutionSettings()
+    root_user: Optional[RootUserSettings] = RootUserSettings()
+    admin_ui: Optional[AdminUiSettings] = AdminUiSettings()
+    notifications: Optional[
+        FidesopsNotificationSettings
+    ] = FidesopsNotificationSettings()
 
-    port: int
+    port: int = int(
+        os.getenv(
+            "FIDESOPS__PORT",
+            "8080",  # Run the webserver on port 8080 by default
+        )
+    )
     is_test_mode: bool = os.getenv("TESTING", "").lower() == "true"
     hot_reloading: bool = os.getenv("FIDESOPS__HOT_RELOAD", "").lower() == "true"
     dev_mode: bool = os.getenv("FIDESOPS__DEV_MODE", "").lower() == "true"